platform: Share more files between stm32 platforms
