+++ app/persistent.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <stdint.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 struct persistent {
 	bool reboot_to_bootloader;
