+++ platforms/stm32_common/pcd.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "usb/def.h"
 #include "errors.h"
+#include "usb/def.h"
 
 err_t pcd_init(USBD_HandleTypeDef *p_usbd);
