+++ platforms/stm32_common/usb/cdc_internal.h
@@ -1,6 +1,7 @@
 #pragma once
 
-#include "stm32l4xx_hal.h"
+#include "stm32_hal.h"
+#include "usb/core.h"
 
 struct cdc_init_data {
 	USBD_HandleTypeDef *p_usbd;
