+++ platforms/stm32_common/usb/core.h
@@ -27,8 +27,8 @@
 
 #pragma once
 
+#include "stm32_hal.h"
 #include "usb/def.h"
-#include "stm32l4xx_hal.h"
 
 HAL_StatusTypeDef USBD_Init(USBD_HandleTypeDef *p_dev, PCD_HandleTypeDef *p_pcd, USBD_DescriptorsTypeDef *p_desc);
 HAL_StatusTypeDef USBD_DeInit(USBD_HandleTypeDef *p_dev);
