+++ platforms/stm32_common/usb/def.h
