+++ platforms/stm32f0xx/stm32_hal.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "stm32f0xx_hal.h"
