+++ platforms/stm32l4xx/gpio.c
@@ -1,5 +1,4 @@
 #include "platform/gpio.h"
-
 #include "stm32l4xx_hal.h"
 
 void gpio_init(void)
