--- platforms/stm32l4xx/pcd.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "usb/def.h"
-#include "errors.h"
-
-err_t pcd_init(USBD_HandleTypeDef *p_usbd);
