+++ platforms/stm32l4xx/stm32_hal.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "stm32l4xx_hal.h"
